Redis container 'chowns' its entire data dir to a 'redis' user inside the container - so we can't map this container to a new user
The redis container runs as the user 'systemd-coredump' on the host on both Ubuntu and CentOS. Since there is a user collision mapping, we would like to change the redis container to use a user that isn't the same user as a systemd component, which might have security implications somewhere down the line. However, the redis container makes this impossible, since it changes all the files in the data base to be owned by the redis user inside the container, and not the user running the entrypoint.