Add ability to specify infrastructure images
Minio, Elasticsearch, and Redis all might need to be pulled from a local
registry. This allows that use case.

We don't default to a puppet-owned repo, since minio is GPL code, and
shipping the code from a puppet-hosted environment is a touch risky, so
better safe than sorry for now.

We don't have anything we don't want to opensource, but better safe than
sorry in the short term.