Add a flag that would force a user to specify a ca, cert, and key to use.
Without this flag, the HDP will make an initial untrusted connection to the Puppet ca server in order to download the CA cert. Users can avoid this by specifying the cert, ca, and key they want to use. We add a boolean parameter now that would force a user to specify these, so that we can check at compile time of the module if the untrusted connection will occur.